requirements.txt
Django==3.2.16
djangorestframework==3.12.4
requests==2.26.0
djangorestframework-simplejwt==4.7.2
